Project/gpt-interface-api/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
distro==1.8.0
exceptiongroup==1.1.3
fastapi==0.103.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
multidict==6.0.4
openai==1.2.3
pydantic==2.3.0
pydantic_core==2.6.3
python-decouple==3.8
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2
